+++ crates/membrane-shadow/src/forgejo/mod.rs
@@ -24,7 +24,7 @@ pub(crate) const API_TIMEOUT_READ: std::time::Duration =
     std::time::Duration::from_secs(cellmembrane_types::service::DEFAULT_API_READ_TIMEOUT_SECS);
 const API_TIMEOUT_FAST: std::time::Duration =
     std::time::Duration::from_secs(cellmembrane_types::service::DEFAULT_PROBE_TIMEOUT_SECS);
-const PAGE_SIZE: usize = cellmembrane_types::service::DEFAULT_API_PAGE_SIZE as usize;
+const PAGE_SIZE: usize = cellmembrane_types::service::DEFAULT_API_PAGE_SIZE;
 
 // ── Types ───────────────────────────────────────────────────────────
